Reworked tests to work with webpack, moved storage interactions to separate file
