+++ package.json
@@ -6,14 +6,15 @@
     "start": "webpack --watch --config webpack.dev.js",
     "build": "webpack --config webpack.prod.js",
     "lint": "eslint --ext .ts,.js --max-warnings=0 .",
-    "prettier": "prettier --write"
+    "prettier": "prettier --write",
+    "test": "nyc --reporter=lcov --reporter=text mocha --require @babel/register"
   },
-  "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.18.6",
+    "@babel/core": "^7.21.8",
     "@babel/plugin-transform-runtime": "^7.18.6",
-    "@babel/preset-env": "^7.18.6",
+    "@babel/preset-env": "^7.21.5",
     "@babel/preset-typescript": "^7.18.6",
+    "@babel/register": "^7.21.0",
     "@types/chrome": "^0.0.193",
     "@typescript-eslint/eslint-plugin": "^5.30.5",
     "@typescript-eslint/parser": "^5.30.5",
@@ -24,16 +25,23 @@
     "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-webpack-plugin": "^3.2.0",
+    "expect.js": "^0.3.1",
     "husky": "^8.0.1",
+    "jsdom": "^22.0.0",
     "lint-staged": "^13.0.3",
     "mini-css-extract-plugin": "^2.6.1",
+    "mocha": "^10.2.0",
+    "nyc": "^15.1.0",
     "prettier": "^2.7.1",
     "sass": "^1.53.0",
     "sass-loader": "^13.0.2",
+    "sinon": "^15.1.0",
+    "sinon-chrome": "^3.0.1",
     "typescript": "^4.7.4",
     "webpack": "^5.73.0",
     "webpack-cli": "^4.10.0",
-    "webpack-merge": "^5.8.0"
+    "webpack-merge": "^5.8.0",
+    "rewire": "^6.0.0"
   },
   "dependencies": {
     "@babel/runtime": "^7.18.6",
