+++ test/utils.test.js
@@ -4,30 +4,11 @@ const rewire = require('rewire');
 const jsdom = require('jsdom');
 const { JSDOM } = jsdom;
 
-const testUtils = rewire('./testUtils.js');
+// const testUtils = rewire('./testUtils.js');
 
-const utils = rewire('../extension/js/utils.js');
+const utils = rewire('../src/utils.js');
 
 describe('utils.js', function () {
-
-	let mockChrome, setupMockSyncStorageObject, setupMockLocalStorageObject;
-
-	this.beforeAll(function () {
-		mockChrome = testUtils.__get__('mockChrome');
-
-		global.configSync = {}, global.mockLocalStorageObject = {};
-		global.chrome = mockChrome();
-
-		setupMockSyncStorageObject = testUtils.__get__('setupMockSyncStorageObject');
-		setupMockLocalStorageObject = testUtils.__get__('setupMockLocalStorageObject');
-	});
-
-	// Restore everything
-	this.afterAll(function () {
-		delete global.chrome;
-		delete global.configSync;
-		delete global.mockLocalStorageObject;
-	});
 
 	context('console helpers', function () {
 
@@ -65,49 +46,6 @@ describe('utils.js', function () {
 			dom = new JSDOM(`<!DOCTYPE html><body><span id="test-span"></span></body>`);
 			dom.window.document.getElementById("test-span").innerText = "Before";
 		});
-
-		context('setDOMTextWithDelay()', function () {
-			const setDOMTextWithDelay = utils.__get__('setDOMTextWithDelay');
-
-			it('should replace DOM text with default predicate', async function () {
-				expect(dom.window.document.getElementById("test-span").innerText).to.be("Before");
-
-				setDOMTextWithDelay(dom.window.document.getElementById("test-span"), "After", 30);
-
-				await new Promise(r => setTimeout(r, 10));
-				expect(dom.window.document.getElementById("test-span").innerText).to.be("Before");
-
-				await new Promise(r => setTimeout(r, 30));
-				expect(dom.window.document.getElementById("test-span").innerText).to.be("After");
-			});
-
-			it('should replace DOM text if custom predicate is true', async function () {
-				const someBoolean = true;
-				const predicate = () => { return dom.window.document.getElementById("test-span").innerText === "Before" && someBoolean; };
-
-				expect(dom.window.document.getElementById("test-span").innerText).to.be("Before");
-
-				setDOMTextWithDelay(dom.window.document.getElementById("test-span"), "After", 30, predicate);
-
-				await new Promise(r => setTimeout(r, 10));
-				expect(dom.window.document.getElementById("test-span").innerText).to.be("Before");
-
-				await new Promise(r => setTimeout(r, 30));
-				expect(dom.window.document.getElementById("test-span").innerText).to.be("After");
-			});
-
-			it('should not replace DOM text if predicate is false', async function () {
-				const someBoolean = false;
-				const predicate = () => { return dom.window.document.getElementById("test-span").innerText === "Before" && someBoolean; };
-
-				expect(dom.window.document.getElementById("test-span").innerText).to.be("Before");
-
-				setDOMTextWithDelay(dom.window.document.getElementById("test-span"), "After", 20, predicate);
-
-				await new Promise(r => setTimeout(r, 30));
-				expect(dom.window.document.getElementById("test-span").innerText).to.be("Before");
-			});
-		});
 	});
 
 	context('URL helpers', function () {
@@ -131,58 +69,6 @@ describe('utils.js', function () {
 				expect(isVideoUrl("https://www.youtube.com")).to.be(false);
 				expect(isVideoUrl("https://www.youtube.com/playlist?list=PL1234567890")).to.be(false);
 			});
-
-		});
-
-		context('loadJsonFile()', function () {
-			const loadJsonFile = utils.__get__('loadJsonFile');
-
-			// Mock the XMLHttpRequest object
-			let xhr, requests;
-			beforeEach(function () {
-				// Replace XMLHttpRequest with FakeXMLHttpRequest
-				xhr = sinon.useFakeXMLHttpRequest();
-				global.XMLHttpRequest = xhr;
-
-				requests = [];
-
-				xhr.onCreate = function (xhr) {
-					requests.push(xhr);
-				};
-			});
-
-			afterEach(function () {
-				// Restore original XMLHttpRequest
-				xhr.restore();
-			});
-
-			it('should load a JSON file', async function () {
-				let json = loadJsonFile("test/test.json");
-
-				// Fake the response
-				requests[0].respond(200, { "Content-Type": "application/json" }, '{ "test": "test" }');
-
-				json = await json;
-
-				expect(json).to.be.an('object');
-				expect(json.test).to.be("test");
-			});
-
-			it('should throw an error if the file is not found', async function () {
-				let json = loadJsonFile("test/doesNotExist.json");
-
-				// Fake the response
-				requests[0].respond(404);
-
-				try {
-					json = await json;
-					expect(true).to.be(false);
-				} catch (e) {
-					expect(e.message).to.be("Not Found");
-				}
-
-			}
-			);
 
 		});
 	});
@@ -246,135 +132,6 @@ describe('utils.js', function () {
 				expect(date.toISOString()).to.be("2018-12-31T23:00:00.000Z");
 			});
 		});
-
-	});
-
-	context('browser storage', function () {
-		const fetchConfigSync = utils.__get__('fetchConfigSync');
-		const setSyncStorageValue = utils.__get__('setSyncStorageValue');
-		const addHours = utils.__get__('addHours');
-
-		this.beforeEach(async function () {
-			await setupMockSyncStorageObject();
-			await setupMockLocalStorageObject();
-		});
-
-		context('fetchConfigSync()', function () {
-
-			it('should return the correct config', async function () {
-				let config = await fetchConfigSync();
-
-				// Only test for some properties, a full test is done separately
-				expect(config).to.be.an('object');
-				expect(config).to.have.property('shuffleOpenInNewTabOption');
-				expect(config.customYoutubeApiKey).to.be(null);
-				expect(config.channelSettings).to.be.an('object');
-				expect(config.channelSettings).to.be.empty;
-			});
-		});
-
-		context('setSyncStorageValue()', function () {
-
-			this.beforeAll(async function () {
-				sinon.spy(chrome.runtime, "sendMessage");
-				sinon.spy(chrome.storage.sync, "set");
-			});
-
-			this.afterAll(async function () {
-				chrome.runtime.sendMessage.restore();
-				chrome.storage.sync.set.restore();
-			});
-
-			this.beforeEach(async function () {
-				// Reset the spies
-				chrome.runtime.sendMessage.resetHistory();
-				chrome.storage.sync.set.resetHistory();
-			});
-
-			it('should set the correct value in the global config object', async function () {
-				await setSyncStorageValue("testAddedKeyGlobal", "testAddedValGlobal");
-
-				expect(configSync.testAddedKeyGlobal).to.be("testAddedValGlobal");
-				expect(chrome.runtime.sendMessage.calledOnce).to.be(true);
-				expect(chrome.storage.sync.set.calledOnce).to.be(true);
-				expect(chrome.runtime.sendMessage.calledWith({ command: "newConfigSync", data: configSync })).to.be(true);
-
-				expect(chrome.runtime.sendMessage.returnValues[0]).to.equal("New configSync set.");
-			});
-
-			it('should set the correct value in the passed config object', async function () {
-				let passedConfigSync = {};
-				await setSyncStorageValue("testAddedKeyPassed", "testAddedValPassed", passedConfigSync);
-
-				expect(passedConfigSync.testAddedKeyPassed).to.be("testAddedValPassed");
-				expect(chrome.runtime.sendMessage.calledOnce).to.be(true);
-				expect(chrome.storage.sync.set.calledOnce).to.be(true);
-
-				expect(chrome.runtime.sendMessage.calledWith({ command: "newConfigSync", data: configSync })).to.be(true);
-			});
-
-			it('should correctly overwrite the global configSync object with the new one', async function () {
-				let passedConfigSync = {};
-				// Remember what the configSync was before to make sure it was replaced
-				let configSyncBefore = configSync;
-
-				await setSyncStorageValue("testAddedKeyPassed", "testAddedValPassed", passedConfigSync);
-
-				expect(passedConfigSync.testAddedKeyPassed).to.be("testAddedValPassed");
-				expect(chrome.runtime.sendMessage.calledOnce).to.be(true);
-				expect(chrome.storage.sync.set.calledOnce).to.be(true);
-
-				expect(chrome.runtime.sendMessage.calledWith({ command: "newConfigSync", data: passedConfigSync })).to.be(true);
-				expect(chrome.runtime.sendMessage.calledWith({ command: "newConfigSync", data: configSyncBefore })).to.be(false);
-				// Make sure the global configSync object was replaced
-				expect(configSync).to.equal(passedConfigSync);
-			});
-
-		});
-
-		context('getUserQuotaRemainingToday()', function () {
-			const getUserQuotaRemainingToday = utils.__get__('getUserQuotaRemainingToday');
-
-			this.beforeAll(async function () {
-				sinon.spy(chrome.storage.sync, "set");
-			});
-
-			this.afterAll(async function () {
-				chrome.storage.sync.set.restore();
-			});
-
-			this.beforeEach(async function () {
-				// Reset the spy
-				chrome.storage.sync.set.resetHistory();
-			});
-
-			it('should return the correct value', async function () {
-				// Set the quota to 100
-				await setSyncStorageValue("userQuotaRemainingToday", 100);
-
-				let quota = await getUserQuotaRemainingToday(configSync);
-
-				expect(quota).to.be(100);
-				expect(chrome.storage.sync.set.calledOnce).to.be(true);
-			});
-
-			it('should correctly reset the quota if the reset time has passed', async function () {
-				// Set the quota to 100
-				await setSyncStorageValue("userQuotaRemainingToday", 100);
-				// Set the reset time to 1 hour ago
-				await setSyncStorageValue("userQuotaResetTime", addHours(new Date(), -1).getTime());
-
-				expect(configSync.userQuotaRemainingToday).to.be(100);
-				expect(configSync.userQuotaResetTime).to.be.lessThan(new Date().getTime());
-
-				let quota = await getUserQuotaRemainingToday(configSync);
-
-				expect(quota).to.be(200);
-				expect(chrome.storage.sync.set.callCount).to.be(4);
-				expect(chrome.storage.sync.set.calledWith({ "userQuotaRemainingToday": 200 })).to.be(true);
-				expect(chrome.storage.sync.set.calledWith({ "userQuotaResetTime": new Date(new Date().setHours(24, 0, 0, 0)).getTime() })).to.be(true);
-			});
-		});
 	});
 
 	context('custom errors', function () {
@@ -432,7 +189,7 @@ describe('utils.js', function () {
 			});
 
 			it('should have the correct properties', function () {
-				const e = new YoutubeAPIError(code = "RYV-test", message = 'test message', reason = 'test reason', solveHint = 'test solveHint', showTrace = true);
+				const e = new YoutubeAPIError("RYV-test", 'test message', 'test reason', 'test solveHint', true);
 
 				expect(e).to.have.property('code');
 				expect(e).to.have.property('message');
